データの読み込みを1つのフックに統合
